study_agents/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
chromadb==0.4.22
pypdf==4.0.1
python-dotenv==1.0.0
openai==1.12.0
tiktoken==0.5.2
faiss-cpu==1.7.4
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

